Updated App to AndroidX
Updated the app to target API 28 (Android P). Also, changed the external library selection from Maven to AndroidX (as it is required with API 28). Created the Azalea Belles and the 2020 Azalea Gardens Map page adding a description of the Belles and a Google Map with points of interest respectively. Also, changed the title of one of the buttons on the 2020 Azalea Garden Tour page.